+++ CHANGELOG.md
@@ -1,66 +1,81 @@
 # Changelog
 
-## 2.1.0
-- Updated: layout and from
-- Updated: to include route by default without config
-
 ## 2.0.0
-- Updated for Livewire 3 and Laravel 11
+
+- Updated for Livewire 4 and Laravel 12
 
 ## 1.3.0
+
 - Changed: from to replyTo in message received to prevent 550 error
 
 ## 1.2.5
+
 - Fixed: reply email address when message received
 
 ## 1.2.4
+
 - Updated: JTB classes and layouts
 
 ## 1.2.3
+
 - Add: page title to view and max width
 
 ## 1.2.2
+
 - fix class typo
 
 ## 1.2.1
+
 - update mail classes to new laravel syntax
 
 ## 1.2.0
+
 - add defer to inputs to fix time out errors
 
 ## 1.1.0
+
 - update gotime form components
 - fix publish view
 
 ## 1.0.2
+
 - minor template tweaks
 
 ## 1.0.1
+
 - update control components
 - update layout
 
 ## 1.0.0
+
 - remove contact controller functionality
 - add contact livewire component
 - replace recaptcha v2 with v3
 
 ## 0.2.1
+
 - update spatie/laravel-honeypot
 
 ## 0.2.0
+
 - update to gotime form components
 
 ## 0.1.1
+
 - update layouts
 
 ## 0.0.4
+
 - update formit components
 
 ## 0.0.3
+
 - fix layout
 
 ## 0.0.2
+
 - fix layout
 
 ## 0.0.1
+
 - Initial release